requirements.txt
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.9.0
Jinja2==2.10.1
Pillow==6.2.2
psycopg2==2.8.4
pytz==2018.5
sqlparse==0.3.1
django-ckeditor==6.0.0
https://github.com/darklow/django-suit/tarball/v2

